requirements.txt
click
dill
einops
glfw
imageio
imageio-ffmpeg
matplotlib
ninja
numpy
onnx
onnxscript
open-clip-torch
pillow
psutil
pyopengl
requests
scipy
setuptools
tensorboard
timm
torch-dct
torch-fidelity
torchmetrics
tqdm
transformers==4.50.1
wandb==0.16.6
webdataset